mudassir-assignment/smartStore-moodAnalyzer/pyproject.toml
[project]
name = "smartstore-moodanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mudassirishere", email = "162560457+mudassirishere@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai>=1.107.2",
    "python-dotenv>=1.1.1",
]

[project.scripts]
smartstore-moodanalyzer = "smartstore_moodanalyzer:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
